pyproject.toml
[tool.poetry]
name = "payment-gateway"
version = "0.1.0"
description = "Simple Payment Gateway"
authors = ["Laith Shadeed <laith.shadeed@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
flask-limiter = "^3.5.0"
flask-migrate = "^4.0.5"
flask-talisman = "^1.1.0"
prometheus-flask-exporter = "^0.22.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
